+++ src/felupe/constitution/strain/models/_linear_elastic_plastic_isotropic.py
@@ -18,7 +18,7 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
 
-from ....math import cdya, ddot, dya, identity, sqrt, trace
+from ....math import cdya_ik, ddot, dya, identity, sqrt, trace
 from ...linear_elasticity import lame_converter
 from .._material_strain import MaterialStrain
 
@@ -147,7 +147,7 @@ def linear_elastic_plastic_isotropic_hardening(dε, εn, σn, ζn, λ, μ, σy, 
     # elasticity tensor
     if kwargs["tangent"]:
         dσdε = np.zeros((3, 3, 3, 3, *dε.shape[2:]))
-        dσdε[:] = λ * dya(eye, eye) + 2 * μ * cdya(eye, eye)
+        dσdε[:] = λ * dya(eye, eye) + 2 * μ * cdya_ik(eye, eye)
     else:
         dσdε = None
 
@@ -187,7 +187,10 @@ def linear_elastic_plastic_isotropic_hardening(dε, εn, σn, ζn, λ, μ, σy, 
                 * μ
                 * dγ
                 / norm_s
-                * (2 * μ * (cdya(eye, eye) - 1 / 3 * dya(eye, eye)) - 2 * μ * dya(n, n))
+                * (
+                    2 * μ * (cdya_ik(eye, eye) - 1 / 3 * dya(eye, eye))
+                    - 2 * μ * dya(n, n)
+                )
             )[..., mask]
 
         # update list of state variables